# src/main/webapp/api-docs/openapi.json
{
    "openapi": "3.0.1",
    "servers": [
        { "url": "/msa-api-gateway/gateway", "description": "API Gateway" }
    ],
...
    "paths": {
...
        "/booking/{id}": {
            "get": {
                "tags": [
                    "Booking"
                ],
                "summary": "Obtener reserva por ID",
                "operationId": "getBookingById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva encontrada"
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    }
                }
            }
        },
        "/flight/{id}": {
            "get": {
                "tags": [
                    "Vuelos"
                ],
                "summary": "Obtener vuelo por ID",
                "operationId": "getFlightById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador del vuelo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vuelo encontrado"
                    },
                    "404": {
                        "description": "Vuelo no encontrado"
                    }
                }
            }
        },
        "/flight/params": {
            "get": {
                "tags": [
                    "Vuelos"
                ],
                "summary": "Buscar vuelos por parÃ¡metros",
                "operationId": "getFlightByParams",
                "parameters": [
                    {
                        "name": "countryOrigin",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "countryDestination",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cityOrigin",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cityDestination",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dateOrigin",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de vuelos obtenida"
                    }
                }
            }
        },
        "/flight/instance/{id}": {
            "get": {
                "tags": [
                    "Vuelos"
                ],
                "summary": "Obtener instancia de vuelo por ID",
                "operationId": "getFlightInstanceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador de la instancia de vuelo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Instancia encontrada"
                    },
                    "404": {
                        "description": "Instancia no encontrada"
                    }
                }
            }
        },
        "/reservation/{id}": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Obtener reserva por ID",
                "operationId": "getReservation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva encontrada"
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    }
                }
            }
        },
        "/room/params": {
            "get": {
                "tags": [
                    "Room"
                ],
                "summary": "Buscar habitaciones por paÃ­s y nombre del hotel",
                "operationId": "getRoomByCountryAndNameHotel",
                "parameters": [
                    {
                        "name": "country",
                        "in": "query",
                        "description": "PaÃ­s donde se ubica el hotel",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nameHotel",
                        "in": "query",
                        "description": "Nombre del hotel",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de habitaciones obtenida"
                    }
                }
            }
        },
        "/room/{id}": {
            "get": {
                "tags": [
                    "Room"
                ],
                "summary": "Obtener una habitaciÃ³n por ID",
                "operationId": "getRoomById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador de la habitaciÃ³n",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HabitaciÃ³n encontrada"
                    },
                    "404": {
                        "description": "HabitaciÃ³n no encontrada"
                    }
                }
            }
        },
        "/user/create": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Crear usuario",
                "operationId": "createUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado"
                    }
                }
            }
        },
        "/user": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Obtener usuario por Email y Password",
                "operationId": "getUserById",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginUserDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateUserDTO": {
                "type": "object",
                "properties": {
                    "born": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "surname": {
                        "type": "string"
                    },
                    "typeUser": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "LoginUserDTO": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "description": "Correo electrÃ³nico del usuario",
                        "example": "user@example.com"
                    },
                    "password": {
                        "type": "string",
                        "description": "ContraseÃ±a del usuario",
                        "example": "password123"
                    }
                },
                "description": "Credenciales de acceso del usuario"
            }
        }
    }
}